refactor(prof): Introduce profiling backend abstraction
Split the profiling API from its implementation. The public API
(prof.c) becomes a wrapper, while the current implementation
is moved to prof_base.c. Introduce a backend interface so different
profiling backends can be plugged in without changing the API.
